backend/requirements.txt
requests
beautifulsoup4
google-generativeai
supabase
python-dotenv
Pillow
feedparser
tweepy
python-Levenshtein
